requirements.txt
fastapi
SQLAlchemy
mysql-connector-python
passlib
uvicorn
python-jose
pydantic
python-multipart
bcrypt
alembic
psycopg2-binary
python-dotenv 